grid size from local storage
